Update: ImGui + Scripts
Users can add draw_debug functions to their classes (and bind them in _bind_methods) to be able to draw to the debug menu.

Drawing a scene hierarchy and the user can select nodes in the hierarchy. When a node is selected, if the node implements a "draw_debug" function, it will draw a debug menu for that node with whatever ImGui elements the user wants.

Tweaks to the build scripts so I can use configuration directly for game scons and only have to correct things for godot_cpp/godot.