Update: Web Compile
Been trying for the last couple of days to get web builds compiling for the game project but it keeps failing. I keep seeing errors like: "memory.editor.dev.wasm32.o was not found" (and for other source files). I've tried installing/activating various other emsdk versions, none of them work. I've tried doing a full scons -c (with appropriate args) and that doesn't work. So I'm going to go with the brute force option and try a "hard reset" and just use a clean godot-cpp submodule and see if that helps compilation.

Suppose it could be entirely possible that with my build script the extension.json file contains data that might be throwing the build off somewhere (not sure why or how this would be the case but you never know!).