Update: Web Platform
Web now compiles and I can open up the godot 4.4 web editor and load my project (using the zip folder created during the build step). Although, gdextensions are not supported as of yet according to the official documentation - this is ready for when it is. Web opens up a file browser in the default system browser and allows the user to navigate the bin folder (using serve.py).

You can open up the editor by going to: .web_zip -> godot.editor.html
Once it's loaded in the browser, you can select "browse" for the project directory option and select the zip folder in the local project directory to use and then click "start godot editor". There will be errors that print because gdextension isn't supported yet.