Update: Web Editor Game
Improved the workflow for editor/editor game web platform. web editor will now directly open the editor instead of going to the directory browser. web editor game will run the editor_game export directly in the browser. These can take a while for the initial builds but then after that, they are not that long to iterate on. Will look at implementing a similar workflow for android next.