Update: CI
Should hopefully fix the "warning LNK4099: PDB '' was not found with 'imgui.windows.template_debug.dev.double.x86_64.obj' or at ''; linking object as if no debug info" warnings. Not exactly sure what's happening here, from my understanding this seems to imply that the original file (imgui.h I think?) was compiled with debug symbols (which should be correct because I include debug symbols in the game project) but the cache returns an obj file that doesn't have the debug symbols compiled in?

Either way, I'm not interested in debugging through imgui.h for the purposes of this repository. So I'm including this linker ignore warning here to try and work around this situation. I'd rather the builds all compiled and exported successfully for now.