Update build.py
Need to build some kind of native os build of the godot engine to do a full build. Needed for generating cpp bindings locally and for exporting the game. Web is the first platform that isn't it's own os on github actions so that's why I'm only seeing these issues now.