Android Custom Export Template
Added simple print files function for convenience to help debug CI problems in the future with missing files or incorrect file names (defaults to the current working directory).

Think this is the fix needed for the android export template for template_debug config. No debug symbols on CI since the engine symbols would be too large, so it might just end up using android_release as the name instead.