Update export.py
Attempting to see if importing the project manually before performing an export fixes anything. Commented out the previous code for altering the gdextension file, shouldn't be necessary to do this if the game can be exported correctly with the correct feature flags enabled. This seems to get through locally, it still reports the shader error (as expected) but still finishes the export process.