Update build.py
Make an editor build of the game project instead of template_debug. Locally when exporting for web template_release it seems to export fine with no errors. Not sure if this will "just work" now on CI so I'm committing just this change to see if it makes a difference. I think there's perhaps some issue with opening the editor on windows and it expecting some kind of windows editor dll during the first file scan of the export even though it's ultimately exporting for web.