Update: CI
Have to setup the ios builds like the macos builds because I always need some kind of editor executable since I'm always building the engine from source. So need to install vulkan for ios builds too in this case.

Reorganized order of builds so they match how they display in the github actions viewer (and are consistent with the order godot shows them in too).