Update export.py
Printing debug for gdextension file to make sure it looks how it should for CI. Extra debug printing when the export command fails to show what files are available.

iOS exports are reporting an error that the custom export templates aren't a part of the presets. So I need to see what's in the gdextension file and what's in the export_presets file.