Update: iOS Exports
Seems like there is no way to export for iOS for free, you need to be a part of the apple developer program and provide the appropriate IDs/provisioning UUIDs in the export settings. It uses an xcode project and that build fails if you try to just use a "random" app team ID. So commenting out the export step for iOS builds, can still compile to get coverage but means there won't be automated builds for them. Also removing the signing step since that's been removed in the parent repo too.