Update: Data Validation + Formatting
Formatted all of the game code source files (with the supplied .editorconfig - think I'm just using the godot provided one?)

Investigated into data validation and how to implement it using godot. Found this function: _validate_property which looks like you can just define it and godot will know to call this function when a scene is saved with this object inside of it. Then you can just check to see if the value of the variable is what you would expect and then use WARN_PRINT to write a warning to the editor if the value is something other than expected. Will be useful to make sure exposed variables in the editor are within reason. Basically was looking for some kind of way to implement IsDataValid from UE and this seems like the simplest solution.